init cuda fft
